Refactor date code into common
