+++ provider/cldr/src/transform/dates/mod.rs
@@ -2,6 +2,7 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
+pub(self) mod common;
 pub mod patterns;
 pub mod skeletons;
 pub mod symbols;
